% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regex special characters}
\usage{
escape_regex(pattern)
}
\arguments{
\item{pattern}{Pattern(s) to escape}
}
\value{
Escaped pattern(s)
}
\description{
Function that takes any pattern and escapes special characters so
they can be used in regex as a regular symbols
}
