% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unr_add_js.R
\name{unr_add_js}
\alias{unr_add_js}
\title{Append knitted JS code}
\usage{
unr_add_js(code, name, options)
}
\arguments{
\item{code}{Knitted chunks of code}

\item{name}{Name of the caller}

\item{options}{Chunk options}
}
\value{
knitted JS code, appended to the original code
}
\description{
This function appends knitted javascript (JS) code to the end of the original
chunk code(s). This is necessary to relocate and remove newly created
column layouts.
}
\section{Main idea behind JS}{


After the chunks are knitted and inserted using \link{unr_chunk}, they create a nested div
structure within the main chunk. The number of newly created chunks is
\emph{n} -> based on the number of breaks + 1. This JS code
\enumerate{
\item generates N slides, where only the i-th element is kept,
\item moves the slides after the originaly created slide.
\item removes the original slide, and
\item refreshes Reveal.js.
}

Additional chunk parameters that are handled using javascript are also
parsed. Full list can be found inside \link{unr_code_options}.
}

