% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unr_chunk.R
\name{unr_chunk}
\alias{unr_chunk}
\title{Unriddle a single chunk}
\usage{
unr_chunk(code, name)
}
\arguments{
\item{code}{Vector of codes}

\item{name}{Name of the caller}
}
\value{
List of formatted code chunks
}
\description{
Given a vector of code lines, return a list where each item represents the
formatted code using \link{unr_chunk_code} split by #break. Parameters
\strong{echo=TRUE} and \strong{output-location: column} are also appended.
}
