% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unr_code_options_child.R
\name{unr_code_options_child}
\alias{unr_code_options_child}
\title{Style individual slides}
\usage{
unr_code_options_child(options)
}
\arguments{
\item{options}{Chunk options}
}
\value{
Vector of javascript code
}
\description{
Used to style individual styles (children of the main chunk).
}
