% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unriddle.R
\name{unriddle}
\alias{unriddle}
\title{Hook function}
\usage{
unriddle(before, options, envir, name, ...)
}
\arguments{
\item{before}{Is the code run before or after the chunk?}

\item{options}{Chunk options}

\item{envir}{Chunk environment}

\item{name}{Name of tha parameter (should be unriddle)}

\item{...}{Additional params}
}
\value{
None
}
\description{
Function that is attached to the \link[knitr:knit_hooks]{knitr::knit_hooks} object whenever
the unriddle parameter is set.
}
